+++ aiaccel/hpo/apps/optimize.py
@@ -1,7 +1,6 @@
 from typing import Any
 
 import argparse
-from collections.abc import Callable
 from concurrent.futures import FIRST_COMPLETED, ThreadPoolExecutor, wait
 from importlib import resources
 import json
@@ -15,52 +14,7 @@ from omegaconf import OmegaConf as oc  # noqa: N813
 from optuna.trial import Trial
 
 from aiaccel.config import load_config, print_config, resolve_inherit
-from aiaccel.hpo.optuna.suggest_wrapper import Const, Suggest, SuggestFloat, T
-
-
-class HparamsManager:
-    """
-    Manages hyperparameters for optimization.
-    This class allows defining hyperparameters with various types and provides
-    a method to suggest hyperparameters for a given trial.
-    Attributes:
-        params (dict): A dictionary where keys are hyperparameter names and values
-                       are callables that take a Trial object and return a hyperparameter value.
-    Methods:
-        __init__(**params_def: dict[str, int | float | str | list[int | float] | Suggest[T]]) -> None:
-            Initializes the HparamsManager with the given hyperparameter definitions.
-        suggest_hparams(trial: Trial) -> dict[str, float | int | str | list[float | int | str]]:
-            Suggests hyperparameters for the given trial.
-    """
-
-    def __init__(self, **params_def: dict[str, int | float | str | list[int | float] | Suggest[T]]) -> None:
-        self.params: dict[str, Callable[[Trial], Any]] = {}
-        for name, param in params_def.items():
-            if callable(param):
-                self.params[name] = param
-            else:
-                if isinstance(param, list):
-                    low, high = param
-                    self.params[name] = SuggestFloat(name=name, low=low, high=high)
-                else:
-                    self.params[name] = Const(name=name, value=param)
-
-    def suggest_hparams(self, trial: Trial) -> dict[str, float | int | str | list[float | int | str]]:
-        """
-        Suggests hyperparameters for a given trial.
-        This method generates a dictionary of hyperparameters by applying the
-        parameter functions stored in `self.params` to the provided trial.
-        Args:
-            trial (Trial): An Optuna trial object used to suggest hyperparameters.
-        Returns:
-            dict[str, float | int | str | list[float | int | str]]: A dictionary
-            where keys are parameter names and values are the suggested
-            hyperparameters, which can be of type float, int, str, or a list of
-            these types.
-        """
-
-        return {name: param_fn(trial) for name, param_fn in self.params.items()}
-
+from aiaccel.hpo.optuna.hparams_manager import HparamsManager
 
 def main() -> None:
     """
@@ -102,7 +56,7 @@ def main() -> None:
 
         params:
           _convert_: partial
-          _target_: aiaccel.apps.optimize.HparamsManager
+          _target_: aiaccel.hpo.optuna.hparams_manager.HparamsManager
           x1: [0, 1]
           x2:
             _target_: aiaccel.apps.optimize.SuggestFloat
