+++ examples/hpo/config.yaml
@@ -1,6 +1,6 @@
 params:
   _convert_: partial
-  _target_: aiaccel.hpo.apps.optimize.HparamsManager
+  _target_: aiaccel.hpo.optuna.hparams_manager.HparamsManager
   x1:
     _target_: aiaccel.hpo.optuna.suggest_wrapper.SuggestFloat
     name: x1
