+++ tests/hpo/apps/in_memory.yaml
@@ -9,7 +9,7 @@ study:
 
 params:
   _convert_: partial
-  _target_: aiaccel.hpo.apps.optimize.HparamsManager
+  _target_: aiaccel.hpo.optuna.hparams_manager.HparamsManager
   x1: [0, 1]
   x2:
     _target_: aiaccel.hpo.optuna.suggest_wrapper.SuggestFloat
